requirements.txt
azure-functions
opencv-python-headless
numpy